+++ CHANGES.md
@@ -6,6 +6,8 @@
 
 * Rename `HttpServer::start()` to `HttpServer::run()`
 
+* Allow to gracefully stop test server via `TestServer::stop()`
+
 
 ## [2.0.0-rc] - 2019-12-20